RM # 41812 : Updated hive.sql to change datatype of comm column in emp_ext table

+++ test/hive.sql
@@ -76,7 +76,7 @@ CREATE EXTERNAL TABLE emp_ext (
     mgr               INT,
     hiredate        DATE,
     sal                INT,
-    comm           DECIMAL(7,2),
+    comm           INT,
     deptno          INT
 )
 row format delimited fields terminated by ','
